+++ tests/providers/microsoft/azure/hooks/test_azure_container_volume.py
@@ -19,8 +19,8 @@
 
 import unittest
 
-from airflow.contrib.hooks.azure_container_volume_hook import AzureContainerVolumeHook
 from airflow.models import Connection
+from airflow.providers.microsoft.azure.hooks.azure_container_volume import AzureContainerVolumeHook
 from airflow.utils import db
